refactor(widget): split embed into view components
